reorg tests/debian
